new file:   .htaccess 	modified:   detail_bulan_pembayaran.php 	modified:   index.php 	modified:   pengeluaran.php 	modified:   riwayat.php 	modified:   siswa.php 	deleted:    uang_kas.php 	new file:   uang_kas.sql

+++ pengeluaran.php
@@ -42,36 +42,38 @@
             <h1 class="m-0 text-dark">Pengeluaran</h1>
           </div><!-- /.col -->
           <div class="col-sm text-right">
-            <button class="btn btn-primary" data-toggle="modal" data-target="#tambahPengeluaranModal"><i class="fas fa-fw fa-plus"></i> Tambah Pengeluaran</button>
-            <!-- Modal -->
-            <div class="modal fade text-left" id="tambahPengeluaranModal" tabindex="-1" role="dialog" aria-labelledby="tambahPengeluaranModalLabel" aria-hidden="true">
-              <div class="modal-dialog" role="document">
-                <form method="post">
-                  <div class="modal-content">
-                    <div class="modal-header">
-                      <h5 class="modal-title" id="tambahPengeluaranModalLabel">Tambah Pengeluaran</h5>
-                      <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                      </button>
-                    </div>
-                    <div class="modal-body">
-                      <div class="form-group">
-                        <label for="jumlah_pengeluaran">Jumlah Pengeluaran</label>
-                        <input type="number" name="jumlah_pengeluaran" id="jumlah_pengeluaran" required class="form-control" placeholder="Rp.">
+            <?php if ($_SESSION['id_jabatan'] !== '3'): ?>
+              <button class="btn btn-primary" data-toggle="modal" data-target="#tambahPengeluaranModal"><i class="fas fa-fw fa-plus"></i> Tambah Pengeluaran</button>
+              <!-- Modal -->
+              <div class="modal fade text-left" id="tambahPengeluaranModal" tabindex="-1" role="dialog" aria-labelledby="tambahPengeluaranModalLabel" aria-hidden="true">
+                <div class="modal-dialog" role="document">
+                  <form method="post">
+                    <div class="modal-content">
+                      <div class="modal-header">
+                        <h5 class="modal-title" id="tambahPengeluaranModalLabel">Tambah Pengeluaran</h5>
+                        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                          <span aria-hidden="true">&times;</span>
+                        </button>
                       </div>
-                      <div class="form-group">
-                        <label for="keterangan">Keterangan</label>
-                        <textarea name="keterangan" id="keterangan" required class="form-control"></textarea>
+                      <div class="modal-body">
+                        <div class="form-group">
+                          <label for="jumlah_pengeluaran">Jumlah Pengeluaran</label>
+                          <input type="number" name="jumlah_pengeluaran" id="jumlah_pengeluaran" required class="form-control" placeholder="Rp.">
+                        </div>
+                        <div class="form-group">
+                          <label for="keterangan">Keterangan</label>
+                          <textarea name="keterangan" id="keterangan" required class="form-control"></textarea>
+                        </div>
+                      </div>
+                      <div class="modal-footer">
+                        <button type="button" class="btn btn-danger" data-dismiss="modal"><i class="fas fa-fw fa-times"></i> Close</button>
+                        <button type="submit" name="btnAddPengeluaran" class="btn btn-primary"><i class="fas fa-fw fa-save"></i> Save</button>
                       </div>
                     </div>
-                    <div class="modal-footer">
-                      <button type="button" class="btn btn-danger" data-dismiss="modal"><i class="fas fa-fw fa-times"></i> Close</button>
-                      <button type="submit" name="btnAddPengeluaran" class="btn btn-primary"><i class="fas fa-fw fa-save"></i> Save</button>
-                    </div>
-                  </div>
-                </form>
+                  </form>
+                </div>
               </div>
-            </div>
+            <?php endif ?>
           </div>
         </div><!-- /.row -->
       </div><!-- /.container-fluid -->
@@ -92,7 +94,9 @@
                     <th>Keterangan</th>
                     <th>Tanggal Pengeluaran</th>
                     <th>Jumlah Pengeluaran</th>
-                    <th>Aksi</th>
+                    <?php if ($_SESSION['id_jabatan'] !== '3'): ?>
+                      <th>Aksi</th>
+                    <?php endif ?>
                   </tr>
                 </thead>
                 <tbody>
@@ -104,41 +108,43 @@
                       <td><?= $dp['keterangan']; ?></td>
                       <td><?= date("d-m-Y, H:i:s", $dp['tanggal_pengeluaran']); ?></td>
                       <td>Rp. <?= number_format($dp['jumlah_pengeluaran']); ?></td>
-                      <td>
-                        <a href="" class="badge badge-success" data-toggle="modal" data-target="#editPengeluaranModal<?= $dp['id_pengeluaran']; ?>"><i class="fas fa-fw fa-edit"></i> Ubah</a>
-                        <div class="modal fade text-left" id="editPengeluaranModal<?= $dp['id_pengeluaran']; ?>" tabindex="-1" role="dialog" aria-labelledby="editPengeluaranModalLabel<?= $dp['id_pengeluaran']; ?>" aria-hidden="true">
-                          <div class="modal-dialog" role="document">
-                            <form method="post">
-                              <input type="hidden" name="id_pengeluaran" value="<?= $dp['id_pengeluaran']; ?>">
-                              <div class="modal-content">
-                                <div class="modal-header">
-                                  <h5 class="modal-title" id="editPengeluaranModalLabel<?= $dp['id_pengeluaran']; ?>">Ubah Pengeluaran</h5>
-                                  <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                                    <span aria-hidden="true">&times;</span>
-                                  </button>
-                                </div>
-                                <div class="modal-body">
-                                  <div class="form-group">
-                                    <label for="jumlah_pengeluaran<?= $dp['id_pengeluaran']; ?>">Jumlah Pengeluaran</label>
-                                    <input type="number" name="jumlah_pengeluaran" id="jumlah_pengeluaran<?= $dp['id_pengeluaran']; ?>" required class="form-control" placeholder="Rp." value="<?= $dp['jumlah_pengeluaran']; ?>">
+                      <?php if ($_SESSION['id_jabatan'] !== '3'): ?>
+                        <td>
+                          <a href="" class="badge badge-success" data-toggle="modal" data-target="#editPengeluaranModal<?= $dp['id_pengeluaran']; ?>"><i class="fas fa-fw fa-edit"></i> Ubah</a>
+                          <div class="modal fade text-left" id="editPengeluaranModal<?= $dp['id_pengeluaran']; ?>" tabindex="-1" role="dialog" aria-labelledby="editPengeluaranModalLabel<?= $dp['id_pengeluaran']; ?>" aria-hidden="true">
+                            <div class="modal-dialog" role="document">
+                              <form method="post">
+                                <input type="hidden" name="id_pengeluaran" value="<?= $dp['id_pengeluaran']; ?>">
+                                <div class="modal-content">
+                                  <div class="modal-header">
+                                    <h5 class="modal-title" id="editPengeluaranModalLabel<?= $dp['id_pengeluaran']; ?>">Ubah Pengeluaran</h5>
+                                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                                      <span aria-hidden="true">&times;</span>
+                                    </button>
                                   </div>
-                                  <div class="form-group">
-                                    <label for="keterangan<?= $dp['id_pengeluaran']; ?>">Keterangan</label>
-                                    <textarea name="keterangan" id="keterangan<?= $dp['id_pengeluaran']; ?>" required class="form-control"><?= $dp['keterangan']; ?></textarea>
+                                  <div class="modal-body">
+                                    <div class="form-group">
+                                      <label for="jumlah_pengeluaran<?= $dp['id_pengeluaran']; ?>">Jumlah Pengeluaran</label>
+                                      <input type="number" name="jumlah_pengeluaran" id="jumlah_pengeluaran<?= $dp['id_pengeluaran']; ?>" required class="form-control" placeholder="Rp." value="<?= $dp['jumlah_pengeluaran']; ?>">
+                                    </div>
+                                    <div class="form-group">
+                                      <label for="keterangan<?= $dp['id_pengeluaran']; ?>">Keterangan</label>
+                                      <textarea name="keterangan" id="keterangan<?= $dp['id_pengeluaran']; ?>" required class="form-control"><?= $dp['keterangan']; ?></textarea>
+                                    </div>
+                                  </div>
+                                  <div class="modal-footer">
+                                    <button type="button" class="btn btn-danger" data-dismiss="modal"><i class="fas fa-fw fa-times"></i> Close</button>
+                                    <button type="submit" name="btnEditPengeluaran" class="btn btn-primary"><i class="fas fa-fw fa-save"></i> Save</button>
                                   </div>
                                 </div>
-                                <div class="modal-footer">
-                                  <button type="button" class="btn btn-danger" data-dismiss="modal"><i class="fas fa-fw fa-times"></i> Close</button>
-                                  <button type="submit" name="btnEditPengeluaran" class="btn btn-primary"><i class="fas fa-fw fa-save"></i> Save</button>
-                                </div>
-                              </div>
-                            </form>
+                              </form>
+                            </div>
                           </div>
-                        </div>
-                        <?php if ($_SESSION['id_jabatan'] == '1'): ?>
-                          <a href="hapus_pengeluaran.php?id_pengeluaran=<?= $dp['id_pengeluaran']; ?>" class="badge badge-danger btn-delete" data-nama="Pengeluaran : Rp. <?= number_format($dp['jumlah_pengeluaran']); ?> | <?= $dp['keterangan']; ?>"><i class="fas fa-fw fa-trash"></i> Hapus</a>
-                        <?php endif ?>
-                      </td>
+                          <?php if ($_SESSION['id_jabatan'] == '1'): ?>
+                            <a href="hapus_pengeluaran.php?id_pengeluaran=<?= $dp['id_pengeluaran']; ?>" class="badge badge-danger btn-delete" data-nama="Pengeluaran : Rp. <?= number_format($dp['jumlah_pengeluaran']); ?> | <?= $dp['keterangan']; ?>"><i class="fas fa-fw fa-trash"></i> Hapus</a>
+                          <?php endif ?>
+                        </td>
+                      <?php endif ?>
                     </tr>
                   <?php endforeach ?>
                 </tbody>
